Handle port range parsing for tp_dst and tp_src keys
Parsing of a flow with tp_dst and tp_src in port/mask format does not work.
Currently parseIntMatch() assumes only a single port in tp_dst and tp_src and not a port-range. As a result parsing a flow with port range fails.
This fix adds logic to parse tp_dst and tp_src in port-range format in a flow.

tp_dst=0xea60/0xffe0