ovsnl: only fail if no known families are found (#88)
Before this commit the ovnl client logic was such that it needs to be
kept in lock-step synchronization with a particular version of the
netlink interface. This is too fragile, as each new update requires
touching this logic, which in turn makes it non-backward compatible.

This commit flips the logic on its head and instead makes the client
complains only if no 'known' families are found. This makes the init
logic future proof and backward compatible.